requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
openai==1.12.0
slowapi==0.1.9
orjson==3.9.15
